+++ dormyboba/handlers/__init__.py
@@ -1,3 +1,6 @@
-from .handlers import handlers_labeler
+from .common import common_labeler
+from .invite import invite_labeler
+from .mailing import mailing_labeler
+from .queue import queue_labeler
 
 __all__ = ("handlers_labeler")
